Many updates

+++ scripts/99.optimize-second-users-ui-only.ps1
@@ -9,7 +9,7 @@ Import-Module -DisableNameChecking $PSScriptRoot\..\lib\common-lib.psm1
 # Preparation
 #============================================================================
 
-Question-UserAdminsRights
+Check-AdminsRights
 
 # Real work
 #=============================================================================
@@ -20,10 +20,15 @@ Print-Message-With-Banner("Starting User UI Cleanup")
 &($PSScriptRoot+"\6.1.cleanup-taskbar.ps1")
 &($PSScriptRoot+"\7.unbloat-start-menu.ps1")
 
-$choice = [Microsoft.VisualBasic.Interaction]::MsgBox('Remove current user from Administrators group?', 'YesNo,SystemModal,Question', 'Remove user from admin group check')
+if (Is-BuiltInAdmin) {
+    Return
+} else {
 
-switch  ($choice) {
-'Yes' {
-    Remove-CurrentUserAdminGroup
-	}
+    $choice = [Microsoft.VisualBasic.Interaction]::MsgBox('Remove current user from Administrators group?', 'YesNo,SystemModal,Question', 'Remove user from admin group check')
+
+    switch  ($choice) {
+    'Yes' {
+        Remove-CurrentUserAdminGroup
+	    }
+    }
 }
